extracted book list back to a separate file
